[#13] refacotring for multiple exporters
